pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.9",
    "Flask-SQLAlchemy>=3.1.1",
    "sqlalchemy>=2.0.38",
    "flask-migrate>=4.1.0",
]

[project.scripts]
start = "gunicorn page_analyzer.app:app"
